refactor: convert coordinator to package and extract setup mixin
