+++ custom_components/ecoflow_energy/coordinator/core.py
@@ -20,7 +20,7 @@ import time
 from collections import deque
 from dataclasses import dataclass, field
 from datetime import timedelta
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
@@ -28,15 +28,12 @@ from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
-from .const import (
+from ..const import (
     AUTH_METHOD_APP,
     AUTH_METHOD_DEVELOPER,
-    CONF_ACCESS_KEY,
     CONF_AUTH_METHOD,
     CONF_EMAIL,
     CONF_PASSWORD,
-    CONF_SECRET_KEY,
-    CONF_USER_ID,
     DELTA_ENERGY_FROM_API,
     DELTA_POWER_TO_ENERGY,
     DEVICE_TYPE_DELTA,
@@ -72,26 +69,29 @@ from .const import (
     SMARTPLUG_STALE_THRESHOLD_S,
     get_delta_profile,
 )
-from .ecoflow.cloud_http import EcoFlowHTTPQuota
-from .ecoflow.cloud_mqtt import EcoFlowMQTTClient
-from .ecoflow.energy_integrator import EnergyIntegrator
-from .ecoflow.iot_api import IoTApiClient
-from .ecoflow.parsers.delta import parse_delta_report
-from .ecoflow.parsers.delta_http import parse_delta_http_quota
-from .ecoflow.parsers.delta3_http import parse_delta3_http_quota
-from .ecoflow.parsers.delta3_proto import (
+from ..ecoflow.energy_integrator import EnergyIntegrator
+from ..ecoflow.parsers.delta import parse_delta_report
+from ..ecoflow.parsers.delta_http import parse_delta_http_quota
+from ..ecoflow.parsers.delta3_http import parse_delta3_http_quota
+from ..ecoflow.parsers.delta3_proto import (
     parse_delta3_cms_heartbeat,
     parse_delta3_display_property,
 )
-from .ecoflow.parsers.powerocean_proto import (
+from ..ecoflow.parsers.powerocean_proto import (
     flatten_heartbeat,
     remap_bp_keys,
     remap_proto_keys,
 )
-from .ecoflow.parsers.powerocean import parse_powerocean_http_quota
-from .ecoflow.parsers.smartplug import parse_smartplug_http_quota, parse_smartplug_report
-from .ecoflow.parsers.stream_proto import parse_stream_proto_message
-from .ecoflow.proto.runtime import decode_proto_runtime_frame
+from ..ecoflow.parsers.powerocean import parse_powerocean_http_quota
+from ..ecoflow.parsers.smartplug import parse_smartplug_http_quota, parse_smartplug_report
+from ..ecoflow.parsers.stream_proto import parse_stream_proto_message
+from ..ecoflow.proto.runtime import decode_proto_runtime_frame
+from .setup import SetupMixin
+
+if TYPE_CHECKING:
+    from ..ecoflow.cloud_http import EcoFlowHTTPQuota
+    from ..ecoflow.cloud_mqtt import EcoFlowMQTTClient
+    from ..ecoflow.iot_api import IoTApiClient
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -106,7 +106,10 @@ class DeviceSnapshot:
     key_count: int = 0
 
 
-class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
+class EcoFlowDeviceCoordinator(
+    SetupMixin,
+    DataUpdateCoordinator[dict[str, Any]],
+):
     """Coordinator for a single EcoFlow device."""
 
     config_entry: ConfigEntry
@@ -122,7 +125,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         # Re-classify device type from product_name if stored as "unknown"
         stored_type = device_info.get("device_type", "")
         if stored_type == DEVICE_TYPE_UNKNOWN:
-            from .const import get_device_type
+            from ..const import get_device_type
             stored_type = get_device_type(device_info.get("product_name", ""), self.device_sn)
         self.device_type: str = stored_type
         display_name = DEVICE_TYPE_DISPLAY_NAMES.get(self.device_type, "")
@@ -416,185 +419,6 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         return info
 
     # ------------------------------------------------------------------
-    # Setup / teardown
-    # ------------------------------------------------------------------
-
-    async def async_setup(self) -> None:
-        """Set up the data source for this device."""
-        self._auth_method = self._entry.data.get(CONF_AUTH_METHOD, AUTH_METHOD_DEVELOPER)
-        session = async_get_clientsession(self.hass)
-
-        # Load energy integrator state from disk (non-blocking)
-        await self.hass.async_add_executor_job(self._energy_integrator.load_state)
-
-        if self._auth_method == AUTH_METHOD_APP:
-            await self._setup_app_auth(session)
-        else:
-            await self._setup_developer_auth(session)
-
-    async def _setup_app_auth(self, session: Any) -> None:
-        """Set up using app authentication (email/password, no Developer API keys).
-
-        App-auth always uses WSS MQTT. No HTTP client or IoT API.
-        """
-        from .ecoflow.app_api import AppApiClient
-
-        email = self._entry.data.get(CONF_EMAIL, "")
-        password = self._entry.data.get(CONF_PASSWORD, "")
-
-        if not email or not password:
-            _LOGGER.error("App-auth: missing credentials for %s", self.device_sn)
-            self._entry.async_start_reauth(self.hass)
-            return
-
-        app_api = AppApiClient(session, email, password)
-        if not await app_api.login():
-            _LOGGER.warning("App-auth: login failed for %s - triggering re-authentication", self.device_sn)
-            self._entry.async_start_reauth(self.hass)
-            return
-
-        user_id = app_api.user_id or self._entry.data.get(CONF_USER_ID, "")
-
-        # No IoT API, no HTTP client for app-auth
-        self._iot_api = None
-        self._http_client = None
-
-        # Fetch portal MQTT credentials (AES-decrypted app-* creds)
-        creds = await app_api.get_mqtt_credentials()
-        if creds is None:
-            _LOGGER.error("App-auth: failed to fetch MQTT credentials for %s", self.device_sn)
-            self._entry.async_start_reauth(self.hass)
-            return
-
-        cert_account = creds.get("certificateAccount") or creds.get("userName", "")
-        cert_password = creds.get("certificatePassword") or creds.get("password", "")
-
-        self._mqtt_client = EcoFlowMQTTClient(
-            certificate_account=cert_account,
-            certificate_password=cert_password,
-            device_sn=self.device_sn,
-            message_handler=self._on_mqtt_message,
-            user_id=user_id,
-            wss_mode=True,
-            enhanced_mode=(self._enhanced_mode and self.device_type == DEVICE_TYPE_POWEROCEAN),
-            auth_error_handler=self._on_mqtt_auth_error,
-        )
-
-        self._credential_obtained_ts = time.monotonic()
-        await self.hass.async_add_executor_job(self._start_mqtt)
-
-        if self._enhanced_mode:
-            if self.device_type == DEVICE_TYPE_POWEROCEAN:
-                self._schedule_keepalive()
-            # The Delta 3 generation never answers the quota request - it
-            # pushes its status frame on its own schedule. The request is
-            # kept anyway: it is one small publish every 30 s and it keeps
-            # outbound traffic on the connection, which is what the other
-            # device families rely on to hold the session open.
-            self._schedule_quotas_poll()
-        self._schedule_ping()
-        self._schedule_stale_check()
-        self._schedule_credential_refresh()
-
-        _LOGGER.debug(
-            "App-auth setup complete for %s (enhanced=%s)",
-            self.device_sn, self._enhanced_mode,
-        )
-
-    async def _setup_developer_auth(self, session: Any) -> None:
-        """Set up using Developer API keys (existing flow, unchanged)."""
-        access_key = self._entry.data.get(CONF_ACCESS_KEY)
-        secret_key = self._entry.data.get(CONF_SECRET_KEY)
-
-        if not access_key or not secret_key:
-            _LOGGER.error("Developer API keys missing for %s - triggering re-authentication", self.device_sn)
-            self._entry.async_start_reauth(self.hass)
-            return
-
-        self._iot_api = IoTApiClient(session, access_key, secret_key)
-
-        self._http_client = EcoFlowHTTPQuota(
-            session, access_key, secret_key, self.device_sn,
-        )
-
-        # Standard Mode: HTTP polling is the primary data source.
-        # MQTT is for SET commands only - except Delta and Smart Plug,
-        # which also subscribe to the IoT MQTT /quota topic for
-        # real-time push alongside HTTP polling.
-        subscribe_mqtt = self.device_type in (
-            DEVICE_TYPE_DELTA,
-            DEVICE_TYPE_DELTA3,
-            DEVICE_TYPE_SMARTPLUG,
-            DEVICE_TYPE_STREAM,
-        )
-        creds = await self._iot_api.get_mqtt_credentials()
-        if creds is not None:
-            cert_account = creds.get("certificateAccount", "")
-            cert_password = creds.get("certificatePassword", "")
-            self._mqtt_client = EcoFlowMQTTClient(
-                certificate_account=cert_account,
-                certificate_password=cert_password,
-                device_sn=self.device_sn,
-                message_handler=self._on_mqtt_message,
-                user_id="",
-                wss_mode=False,
-                subscribe_data=subscribe_mqtt,
-                auth_error_handler=(
-                    self._on_mqtt_auth_error if subscribe_mqtt else None
-                ),
-            )
-            self._credential_obtained_ts = time.monotonic()
-            await self.hass.async_add_executor_job(self._start_mqtt)
-        if subscribe_mqtt:
-            _LOGGER.debug(
-                "Standard Mode + MQTT push: HTTP every %ds + MQTT real-time for %s",
-                HTTP_FALLBACK_INTERVAL_S, self.device_sn,
-            )
-        else:
-            _LOGGER.debug(
-                "Standard Mode: HTTP polling every %ds for %s",
-                HTTP_FALLBACK_INTERVAL_S, self.device_sn,
-            )
-
-    def _start_mqtt(self) -> None:
-        """Start the MQTT client (runs in executor thread)."""
-        if self._mqtt_client is None:
-            return
-        if self._mqtt_client.create_client():
-            if self._mqtt_client.connect():
-                self._mqtt_client.start_loop()
-                mode_label = "WSS Enhanced" if self._enhanced_mode else "TCP Standard"
-                _LOGGER.info("MQTT started for %s (%s)", self.device_sn, mode_label)
-                self._log_event("mqtt_connect", mode_label)
-            else:
-                _LOGGER.error("MQTT connect failed for %s", self.device_sn)
-                self._log_event("mqtt_disconnect", "connect failed")
-        else:
-            _LOGGER.error("MQTT client creation failed for %s", self.device_sn)
-            self._log_event("mqtt_disconnect", "client creation failed")
-
-    async def async_shutdown(self) -> None:
-        """Stop the MQTT client and cancel timers."""
-        self._shutdown = True
-        for handle in (
-            self._keepalive_unsub, self._quotas_unsub, self._ping_unsub,
-            self._stale_check_unsub, self._credential_refresh_unsub,
-            self._powerocean_soc_debounce_unsub,
-        ):
-            if handle is not None:
-                handle.cancel()
-        self._keepalive_unsub = None
-        self._quotas_unsub = None
-        self._ping_unsub = None
-        self._stale_check_unsub = None
-        self._credential_refresh_unsub = None
-        if self._mqtt_client is not None:
-            await self.hass.async_add_executor_job(self._mqtt_client.disconnect)
-            self._mqtt_client = None
-        await self.hass.async_add_executor_job(self._energy_integrator.force_flush)
-        await super().async_shutdown()
-
-    # ------------------------------------------------------------------
     # EnergyStreamSwitch keep-alive (Enhanced Mode only)
     # ------------------------------------------------------------------
 
@@ -746,7 +570,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         if auth_method == AUTH_METHOD_APP:
             # App-auth: re-login and re-fetch portal credentials
             session = async_get_clientsession(self.hass)
-            from .ecoflow.app_api import AppApiClient
+            from ..ecoflow.app_api import AppApiClient
 
             email = self._entry.data.get(CONF_EMAIL, "")
             password = self._entry.data.get(CONF_PASSWORD, "")
@@ -839,7 +663,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
 
         if auth_method == AUTH_METHOD_APP:
             session = async_get_clientsession(self.hass)
-            from .ecoflow.app_api import AppApiClient
+            from ..ecoflow.app_api import AppApiClient
 
             email = self._entry.data.get(CONF_EMAIL, "")
             password = self._entry.data.get(CONF_PASSWORD, "")
@@ -913,7 +737,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         A rejection means the user pressed a control and the device did not
         apply it, which is worth a warning. A successful write stays silent.
         """
-        from .ecoflow.delta3_commands import parse_config_write_ack
+        from ..ecoflow.delta3_commands import parse_config_write_ack
 
         ack = parse_config_write_ack(payload)
         if ack is None:
@@ -1083,13 +907,13 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         (EmsChangeReport, connectivity + enum fields) and cmd_id=13
         (EmsParamChangeReport, the app-side surplus mirror).
         """
-        from .ecoflow.proto.decoder import decode_header_message
-        from .ecoflow.parsers.powerocean_proto import remap_bp_keys
+        from ..ecoflow.proto.decoder import decode_header_message
+        from ..ecoflow.parsers.powerocean_proto import remap_bp_keys
 
         headers, _ = decode_header_message(payload)
 
         try:
-            from .ecoflow.proto import ecocharge_pb2 as pb2
+            from ..ecoflow.proto import ecocharge_pb2 as pb2
             from google.protobuf.json_format import MessageToDict
 
             merged: dict[str, Any] = {}
@@ -1151,7 +975,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
             return parse_stream_proto_message(payload)
 
         if headers is None:
-            from .ecoflow.proto.decoder import decode_header_message
+            from ..ecoflow.proto.decoder import decode_header_message
 
             headers, _ = decode_header_message(payload)
         for hdr in headers or []:
@@ -1164,7 +988,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
                 continue
 
             if self.device_type == DEVICE_TYPE_SMARTPLUG:
-                from .ecoflow.parsers.smartplug import parse_smartplug_proto_heartbeat
+                from ..ecoflow.parsers.smartplug import parse_smartplug_proto_heartbeat
                 result = parse_smartplug_proto_heartbeat(pdata)
                 if result:
                     return result
@@ -1414,7 +1238,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
             _LOGGER.warning("Cannot send SoC limits - MQTT not connected (%s)", self.device_sn)
             return False
 
-        from .ecoflow.energy_stream import build_soc_limit_set_payload
+        from ..ecoflow.energy_stream import build_soc_limit_set_payload
 
         payload = build_soc_limit_set_payload(max_charge_soc, min_discharge_soc)
         ok = await self.hass.async_add_executor_job(
@@ -1510,7 +1334,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
             )
             return False
 
-        from .ecoflow.energy_stream import build_powerocean_soc_set_payload
+        from ..ecoflow.energy_stream import build_powerocean_soc_set_payload
 
         payload = build_powerocean_soc_set_payload(
             backup_reserve_pct,
@@ -1547,7 +1371,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
             )
             return False
 
-        from .ecoflow.energy_stream import build_work_mode_set_payload
+        from ..ecoflow.energy_stream import build_work_mode_set_payload
 
         payload = build_work_mode_set_payload(work_mode)
         ok = await self.hass.async_add_executor_job(
@@ -1616,7 +1440,7 @@ class EcoFlowDeviceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
 
     async def _async_send_delta3_set_proto(self, command: dict[str, Any]) -> bool:
         """Write a Delta 3 setting as a ConfigWrite frame on the app channel."""
-        from .ecoflow.delta3_commands import build_proto_command
+        from ..ecoflow.delta3_commands import build_proto_command
 
         params = command.get("params", {})
         if self._mqtt_client is None or not self._mqtt_client.is_connected():
